Create solpoems
SolPoems is an NFT collection on the Solana blockchain consisting of 279 unique handwritten poems. These poems cover a wide range of topics and follow no certain structure. This will be one of the first NFT projects that include land ownership as a perk of being a part of the community.